// src/app/models/workflow.model.ts
export interface Workflow {
  WObjectID: number;
  WObjectInstanceID: number;
  ObjectTypeDesc: string;
  QFCNumber: string;
  ObjectDetails: string;
  WObjectWorkFlowStatusDesc: string;
  TaskSeq?: number;
  SignatoryName?: string;
  SignatoryRole?: string;
  SignatoryEmailAddress?: string;
  WObjectSoTaskStatusDesc?: string;
  SOStatusAssignedDate?: string;
  SOStatusCompletedDate?: string;
}

export interface Workflow {
  WObjectID: number;
  WObjectInstanceID: number;
  ObjectTypeDesc: string;
  QFCNumber: string;
  ObjectDetails: string;
  WObjectWorkFlowStatusDesc: string;
  TaskSeq?: number;
  SignatoryName?: string;
  SignatoryRole?: string;
  SignatoryEmailAddress?: string;
  WObjectSoTaskStatusDesc?: string;
  SOStatusAssignedDate?: string;
  SOStatusCompletedDate?: string;
}